Don't use [ExpandParams] on Deferred.xxxWith()
On Deferred.resolveWith() and Deferred.rejectWith(), jQuery actually expects an array to be passed as the second argument. 

It's not well documented on their website (it's more apparent in the code) but you can see some evidence:

http://api.jquery.com/deferred.resolve/
http://api.jquery.com/deferred.resolveWith/

Resolve is marked as taking "args" with type "Anything". ResolveWith is marked as taking "args" with type "Array".